Updated:
- ASL 20211109
- Bonmin 1.8.8
- CBC 2.10.5
- Couenne 0.5.8

+++ Couenne/CoinUtils/src/CoinFactorization2.cpp
@@ -1,4 +1,4 @@
-/* $Id: CoinFactorization2.cpp 2083 2019-01-06 19:38:09Z unxusr $ */
+/* $Id: CoinFactorization2.cpp 2213 2019-12-19 08:44:16Z stefan $ */
 // Copyright (C) 2002, International Business Machines
 // Corporation and others.  All Rights Reserved.
 // This code is licensed under the terms of the Eclipse Public License (EPL).
@@ -520,9 +520,9 @@ int CoinFactorization::factorDense()
   totalElements_ = full;
 #ifdef COIN_ALIGN_DENSE
   int newSize = full + 8 * numberDense_;
-  newSize += (numberDense_ + 1) / (sizeof(CoinFactorizationDouble) / sizeof(int));
-  newSize += 2 * ((numberDense_ + 3) / (sizeof(CoinFactorizationDouble) / sizeof(short)));
-  newSize += ((numberRows_ + 3) / (sizeof(CoinFactorizationDouble) / sizeof(short)));
+  newSize += (numberDense_ + 1) / static_cast<int>(sizeof(CoinFactorizationDouble) / sizeof(int));
+  newSize += 2 * ((numberDense_ + 3) / static_cast<int>(sizeof(CoinFactorizationDouble) / sizeof(short)));
+  newSize += ((numberRows_ + 3) / static_cast<int>(sizeof(CoinFactorizationDouble) / sizeof(short)));
   // so we can align on 256 byte
   newSize += 32;
   denseArea_ = new double[newSize];
